Fixed integer types including possible truncation bug

+++ include/gpc/bytes.h
@@ -19,7 +19,7 @@
 #include <sstream>
 extern "C" {
 GP_NONNULL_ARGS(3) GP_PRINTF(3, 4)
-int pf_snprintf(char*GP_RESTRICT, size_t, const char*GP_RESTRICT, ...);
+size_t pf_snprintf(char*GP_RESTRICT, size_t, const char*GP_RESTRICT, ...);
 #endif
